% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdata.R
\docType{data}
\name{bulkdata}
\alias{bulkdata}
\title{Bulk RNA-seq data of tumour cell lines of different cancer type.}
\format{
A data frame with 37004 rows and 1387 columns.
}
\source{
\url{https://cellmodelpassports.sanger.ac.uk/downloads}
}
\usage{
bulkdata
}
\description{
This dataset contains TPM normalized expression profile.
}
\keyword{datasets}
