% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_info.R
\docType{data}
\name{drug_info}
\alias{drug_info}
\title{Information of drugs extraced from GDSC2 project}
\format{
A data frame with 295 rows and 4 columns.
}
\source{
\url{https://www.cancerrxgene.org/downloads/drug_data}
}
\usage{
drug_info
}
\description{
This dataset contains target and targeted signaling pathway of various drugs.
}
\keyword{datasets}
