% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdscdata.R
\docType{data}
\name{gdscdata}
\alias{gdscdata}
\title{Pharmacogenomic data of tumour cell lines.}
\format{
A data frame with 196344 rows and 19 columns.
}
\source{
\url{https://www.cancerrxgene.org/downloads/drug_data}
}
\usage{
gdscdata
}
\description{
This dataset contains IC50 and AUC of cell lines to various drugs.
}
\keyword{datasets}
