% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPharmCombo.R
\name{scPharmCombo}
\alias{scPharmCombo}
\title{Identify potential drug combinations}
\usage{
scPharmCombo(object, score, drug = NULL, topN = 1)
}
\arguments{
\item{object}{Seurat object after running scPharmIdentify function}

\item{score}{output of running scPhrmDr}

\item{drug}{name of drug as drug 1. Default:NULL}

\item{topN}{number of drugs at top of Dr list as drug 1. Default:1}
}
\value{
a list consist of data.frame consist of drug combinations and effects value
}
\description{
Identify potential drug combinations
}
\examples{
combo <- scPharm(scPahrmIdentify.result, scPharmDr.result)
}
