% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPharmGenNullDist.R
\encoding{UTF-8}
\name{scPharmGenNullDist}
\alias{scPharmGenNullDist}
\title{Generate the null distribution and calculate the thresholds for labeling sensitive and resistant cells}
\usage{
scPharmGenNullDist(
  object,
  cancer,
  nmcs = 50,
  nfeatures = 200,
  cores = 1,
  features = NULL,
  slot = "data",
  assay = "RNA",
  bulkdata = bulkdata,
  gdscdata = gdscdata
)
}
\arguments{
\item{object}{Seurat object}

\item{cancer}{the TCGA cancer type associated with your cells for constructing null distribution. A character or vector.(eg: BRCA or c('LUAD', 'LUSC')). cancer='pan' means calculating in the context of pan-cancer.}

\item{nmcs}{number of components to compute and store for MCA. Default:50}

\item{nfeatures}{number of genes used to make cell ID}

\item{cores}{number of CPU cores to use. This parameter can only be set to 1 on windows. Default:1}

\item{features}{character vector of feature names for MCA. If not specified all features will be taken.}

\item{slot}{slot of seurat object used to run MCA}

\item{assay}{assay of seurat object used to run MCA}

\item{bulkdata}{file for all cancer cell lines, automatically loaded in scPharm.}

\item{gdscdata}{pharmacological file for all cancer cell lines, automatically loaded in scPharm.}
}
\value{
a list contains the null distribution(NullDist), thresholds for labeling sensitive and resistant cells(threshold_s and threshold_r).
}
\description{
Generate the null distribution and calculate the thresholds for labeling sensitive and resistant cells
}
\examples{
\dontrun{
result <- scPharmIdentify(seurat.object, type = "tissue", cancer = "LUAD")
}
}
