% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPharmIdentify.R
\name{scPharmIdentify}
\alias{scPharmIdentify}
\title{Identify pharmacological cell subpopulations}
\usage{
scPharmIdentify(
  object,
  type,
  cancer,
  drug = NULL,
  nmcs = 50,
  nfeatures = 200,
  cores = 4,
  features = NULL,
  slot = "data",
  assay = "RNA",
  bulk_data = bulkdata,
  GDSC = gdscdata
)
}
\arguments{
\item{object}{Seurat object}

\item{type}{the source of cell, cell line or tumor tissue. Can be set to 'cellline' or 'tissue'. Default:tissue}

\item{cancer}{the cancer type of cell}

\item{drug}{the drug name for identifying. If not specified all drugs from GDSC2 project will be taken.}

\item{nmcs}{number of components to compute and store for MCA, default set to 50}

\item{nfeatures}{number of genes used to make cell ID}

\item{cores}{number of CPU cores to use}

\item{features}{character vector of feature names. If not specified all features will be taken.}

\item{slot}{slot of seurat object used to run MCA}

\item{assay}{assay of seurat object used to run MCA}
}
\value{
a seurat object added new meta data
}
\description{
Identify pharmacological cell subpopulations
}
\examples{
result <- scPharmIdentify(seurat.object, type = "tissue", cancer = "LUAD")
}
