% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPharmDr.R
\name{scPharmDr}
\alias{scPharmDr}
\title{Compute Dr}
\usage{
scPharmDr(object)
}
\arguments{
\item{object}{Seurat object after running scPharmIdentify function}
}
\value{
a data.frame consist of drug information and drug Dr
}
\description{
Compute Dr
}
\examples{
Dr <- scPharmDr(scPharmIdentify.result)
}
