% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPharmDse.R
\name{scPharmDse}
\alias{scPharmDse}
\title{Predict drug effects}
\usage{
scPharmDse(object)
}
\arguments{
\item{object}{Seurat object after running scPharmIdentify function}
}
\value{
a data.frame consist of drug information and drug Dse
}
\description{
Predict drug effects
}
\examples{
Dse <- scPharmDse(scPharmIdentify.result)
}
